Tellco_Project/requirements.txt
psycopg2
python-dotenv
psycopg2-binary
sqlalchemy
pandas
scikit-learn
streamlit
sqlite3